+++ packages/ts/src/basic.ts
@@ -1,5 +1,7 @@
 import type { RuleValidator } from './validation';
 
+export type Classname = string | string[] | Record<string, any>;
+
 export type ValueSource = 'value' | 'field';
 
 export type ValueEditorType =
@@ -32,13 +34,18 @@ export type OptionGroup<Opt extends Option = Option> = {
 
 export type OptionList<Opt extends Option = Option> = Opt[] | OptionGroup<Opt>[];
 
+interface HasOptionalClassName {
+  className?: Classname;
+}
+
 export interface Field<
   FieldName extends string = string,
   OperatorName extends string = string,
   ValueName extends string = string,
   OperatorObj extends Option = Option<OperatorName>,
   ValueObj extends Option = Option<ValueName>
-> extends Option<FieldName> {
+> extends Option<FieldName>,
+    HasOptionalClassName {
   id?: string;
   operators?: OptionList<OperatorObj>;
   valueEditorType?: ValueEditorType | ((operator: string) => ValueEditorType);
@@ -56,6 +63,8 @@ export type WithRequired<T, K extends keyof T> = T & { [P in K]-?: T[P] };
 
 export type Arity = number | 'unary' | 'binary' | 'ternary';
 
-export interface Operator<N extends string = string> extends Option<N> {
+export interface Operator<N extends string = string> extends Option<N>, HasOptionalClassName {
   arity?: Arity;
 }
+
+export interface Combinator<N extends string = string> extends Option<N>, HasOptionalClassName {}